Crud for technologies completed however some bugs persist CRUD for testimonials started, show and delete are done. Testimonial Create and Update require more effort + JS blades and validation and error message done for the above

+++ app/Http/Controllers/ProjectController.php
@@ -117,6 +117,7 @@ class ProjectController extends Controller
             $project->image = App::make('ImageResize')->imageStore($request->image,'ProjectImg');
         }
         $project->save();
+        if($project->technologies != $request->technologies)
         $project->technologies()->detach();
         foreach($request->technologies as $tech)
         {
